move supporting functions of classifier to a separate file. This should resolve loading issue for the web deployment of the app
