+++ app/run.py
@@ -11,7 +11,7 @@ from sqlalchemy import create_engine
 import sys
 
 sys.path.append('../')
-from models.train_classifier import tokenize
+from models.supporting_functions import tokenize
 
 app = Flask(__name__)
